+++ src/views/AuditLaunch/AuditLaunch.vue
@@ -67,6 +67,11 @@
 
                     <audit-enable-screenshot-form
                         v-model="auditConfigurationForm.common.enableScreenshot"/>
+
+					<audit-css-query-form
+						v-model="auditConfigurationForm.common.cssQuery"
+						:is-valid="isCssQueryValid"
+						:has-been-sent="hasTryToLaunch"/>
                 </section>
             </div>
 
@@ -313,6 +318,7 @@ import AuditBreakpointsForm from "./form/AuditBreakpointsForm";
 import AuditWaitTimeForm from "./form/AuditWaitTimeForm";
 import AuditEnableScreenshotForm from "./form/AuditEnableScreenshotForm";
 import AuditBrowserForm from "./form/AuditBrowserForm";
+import AuditCssQueryForm from './form/AuditCssQueryForm.vue';
 import BasicAuthLoginForm from "./form/BasicAuthLoginForm";
 import BasicAuthPasswordForm from "./form/BasicAuthPasswordForm";
 import BasicAuthUrlForm from "./form/BasicAuthUrlForm";
@@ -322,7 +328,8 @@ import BreakpointHelper from "../../helper/breakpointHelper"
 export default {
     name: 'auditLaunch',
     components: {
-        AuditBrowserForm,
+		AuditBrowserForm,
+        AuditCssQueryForm,
         AuditEnableScreenshotForm,
         AuditWaitTimeForm,
         AuditBreakpointsForm,
@@ -384,6 +391,7 @@ export default {
                     breakpoints: [1920],
                     type: "page",
                     enableScreenshot: false,
+                    cssQuery: '',
                     browser: 'firefox',
                     login: '',
                     password: '',
@@ -478,8 +486,10 @@ export default {
                 'WAIT_TIME': this.auditConfigurationForm.common.waitTime,
                 'WEBDRIVER_RESOLUTIONS': this.auditConfigurationForm.common.breakpoints.join(';'),
                 'ENABLE_SCREENSHOT': this.auditConfigurationForm.common.enableScreenshot,
-                'WEBDRIVER_BROWSER': this.auditConfigurationForm.common.browser
-            };
+				'WEBDRIVER_BROWSER': this.auditConfigurationForm.common.browser,
+				'CSS_QUERY_ON_LOADED_PAGE': this.auditConfigurationForm.common.cssQuery
+			};
+
             switch (this.auditConfigurationForm.common.type) {
                 case 'site':
                     parameters['SITE_SEEDS'] = this.auditConfigurationForm.site.seeds.join(';');
@@ -568,12 +578,22 @@ export default {
         },
         isBrowserValid(){
             return this.activeBrowsers.includes(this.auditConfigurationForm.common.browser);
+		},
+		isCssQueryValid() {
+			try{
+				if(this.auditConfigurationForm.common.cssQuery != ''){
+					document.querySelectorAll(this.auditConfigurationForm.common.cssQuery);
+				}
+				return true;
+			}catch(error){
+				return false;
+			}
         },
         isBreakpointsValid() {
             return this.auditConfigurationForm.common.breakpoints.filter(breakpoint => {
                 return ! BreakpointHelper.isBreakpointValid(breakpoint);
             }).length === 0;
-        },
+		},
         //Pages
         checkValidUrl: UrlHelper.checkValidUrl,
         isPageUrlsValid() {
@@ -617,7 +637,8 @@ export default {
                 this.isMainReferenceValid &&
                 this.isBreakpointsValid &&
                 this.isWaitTimeValid &&
-                this.isBrowserValid;
+				this.isBrowserValid &&
+				this.isCssQueryValid;
             switch (this.auditConfigurationForm.common.type) {
                 case 'scenario':
                     result &= this.isSelectedScenarioValid;